pyproject.toml
[project]
name = "notes-app"
version = "0.1.0"
description = "Educational FastAPI Notes App"
readme = "README.md"
authors = [
    { name = "Educational Project", email = "info@relational-databases.erhardt.consulting" }
]
requires-python = ">=3.11"
keywords = ["fastapi", "sqlalchemy", "orm", "educational", "postgresql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core FastAPI framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    
    # Database and ORM
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    
    # Data validation and serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Templates and forms
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    
    # Utilities
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
]

[project.urls]
Homepage = "https://relational-databases.erhardt.consulting/"
Documentation = "https://relational-databases.erhardt.consulting/"
Repository = "https://github.com/educational/notes-app-python"

[project.scripts]
notes-app = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

